Integrate dockerization into CI
This change overhauls the previous scripts for building and running
docker containers as they became obsolete in the meanwhile. The new
scripts are all rewritten in powershell and cover the two variants
of the server (`blazor` and `core`).

Finally, the building and publishing of Docker containers to
DockerHub has been integrated into the continuous integration.