Convert AasxServerWindows project to SDK-style
This patch converts the AasxServerWindows project from legacy (non-SDK)
to SDK style. This makes it easier to build & publish the releases as
well as to manage the dependencies.

The config `src/AasxServerWindows/App.config` seemed unnecessary and
conflicting with the actual versions that `dotnet build` used, so it was
removed. As it seems, it pinned only the library versions.