Base64URL-encoding for fragments
Previously, fragments were either not encoded at all (AML) or URL-encoded (XML, etc.). In order to be more compliant with the latest version of DotAAS Part 2 and facilitate decoding of fragments, all fragments now need to be Base64URL-encoded.

Base64URL-decoding is now done separately and the decoded fragment is passed to the type-specific routines.