Enable recursive browsing by extensions
This enables extensions to indicate that a certain item in the tree represents a 'browsable' item. For such items, the '(Re-)browse with extension' button will be shown that allows to iterate deeper, possibly using another extension.

One example would be a Zip file that contains an XML file. This change allows browsing into the XML file within the Zip file.