Support recursive browsing into files via the tree (#151)
* Include embeddedDataSpecifications (#142)

A small addition to add the definitions and preferred names of concept descriptions to the output of the /aas/(id)/cds/ REST call.

Also fixes a bug when requesting an AAS by ID that was not found.

* Basic infrastructure for service extensions for browsing into files

* Display tree for AML files

* Move logic for rending extension items to IAASServiceExtension

* Display details for AML file nodes

* Display tree and details for ZIP files

* Allow to (re)browse files using a specific extension

* Display tree and details for XML files

* Display fragment and Rest URL in details (currently only aml)

* Extend the XML Rest API fragments to support attributes

* Refactoring

Migrate some common code for AASServiceExtensions to an abstract base class

* Display fragment and Rest URL for XML files

* Display fragment and Rest URL for Zip files

* Fix issues

Edit works again
Deserialization of AAS
Refresh tree for changed AAS

* Create Queries in AASX-Server.md

* Queries in AASX Server

* Delete Queries in AASX-Server.md

Replaced by PDF

* Add queries documentation as .md

* Display download link for files in ZIP containers

* PUT aas with security check for CREATE and PUT separately

* Add object attributes to PUT submodel elements

* Some refactoring

* Enable recursive browsing by extensions

This enables extensions to indicate that a certain item in the tree represents a 'browsable' item. For such items, the '(Re-)browse with extension' button will be shown that allows to iterate deeper, possibly using another extension.

One example would be a Zip file that contains an XML file. This change allows browsing into the XML file within the Zip file.

* Catch a possible extension during file download

Co-authored-by: Coen van Leeuwen <coenvanl@gmail.com>
Co-authored-by: aorzelskiGH <aorzelski@phoenixcontact.com>
Co-authored-by: aorzelskiGH <48956800+aorzelskiGH@users.noreply.github.com>