clean up pipeline scripts and allow usage of semantic versioning (#268)
Update the process for creating versions:

Update the Changelog
Move all current changes to the [Released] section in the changelog.
Determine the new version number based on semantic versioning and add the release date.
Update the Version Configuration
Use the new version number to update the current_version.cfg.
Push Changes
Push these changes to the branch you want to release from (main/develop/or your branch).
Create a New Version Tag on GitHub
Go to GitHub and create a new version tag on the branch. This tag should be the full version string, such as "0.3.0-1-aasV3-alpha-develop". The build number can be omitted in the tag name.